feat(scheduler): add event loop architecture
We have chosen an event loop architecture approach for inter-thread communication. Once we received a gRPC call from one of the services, it will be sent to the Manager thread through an mpsc channel, procced and the result are sent through a oneshoot channel.

We have chosen this architecture for the following reasons :
- Handle multiple concurrent events processing
- Avoid launching a new thread for each event received
- Examples of mspc channels seen with @sameo

Signed-off-by: iverly <github@iverly.net>